+++ test/tests.jl
@@ -52,6 +52,7 @@ for (test_name, data_norm) in [("mi", data_bin), ("mi_nz", data_mi_nz),
     end
 end
 
+
 # reproduce test results
 # exp_dict = Dict{String, Any}()
 # for (test_name, data_norm) in [("mi", data_bin), ("mi_nz", data_mi_nz),