+++ NEWS.md
@@ -1,3 +1,4 @@
 # Latest
 
-- make Travis CI handle osx (fix HDF5)
+- make Travis CI handle osx (fix HDF5 compilation)
+- remove several unmaintained features (temporary output, AND rule, pvalue weights)